debugging scripted deploy on virtual machine. restructured bash scripts.

+++ programs/python.sh
@@ -1,7 +1,10 @@
 # Install python (stripped down anaconda version)
-wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh 
-bash Miniconda3-latest-Linux-x86_64.sh -b 
-echo 'export PATH="/root/miniconda3/bin:$PATH"' | cat - ~/.bashrc > /tmp/out 
-mv /tmp/out ~/.bashrc 
-rm Miniconda3-latest-Linux-x86_64.sh
-export PATH=~/miniconda3/bin:$PATH
+echo "Installing python..."
+if [[ -d /home/${USER}/miniconda* ]]; then
+  wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh 
+  sudo bash Miniconda3-latest-Linux-x86_64.sh -ub 
+  echo 'export PATH="~/miniconda3/bin:$PATH"' | cat - ~/.bashrc > /tmp/out 
+  mv /tmp/out ~/.bashrc 
+  rm Miniconda3-latest-Linux-x86_64.sh
+  export PATH=~/miniconda3/bin:$PATH
+fi
